Split widget: collections header
